+++ App.tsx
@@ -7,9 +7,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 
-import { ProjectsScreen } from "./src/screens/ProjectsScreen";
-import { SingleProject } from "./src/screens/Singleproject";
-import { DetailsProjectScreen } from "./src/screens/DetailScreen";
+import { AddForm, ProjectsScreen, SingleProject, DetailsProjectScreen } from "./src/screens";
 
 const Stack = createNativeStackNavigator(); 
 const store = createStore(reducer);
@@ -25,6 +23,7 @@ export default function App() {
         <Stack.Screen name="Home" component={ProjectsScreen} />
         <Stack.Screen name="SingleProject" component={SingleProject} />
         <Stack.Screen name="Hart Estimate" component={DetailsProjectScreen} />
+        <Stack.Screen name="Add Details" component={AddForm} />
 
         </Stack.Navigator>
