+++ src/design/Button/index.tsx
@@ -2,22 +2,6 @@ import { FC } from "react";
 import { Button as RNEButton } from "react-native-elements";
 import { StyleSheet } from "react-native";
 
-const styles = StyleSheet.create({
-  titleStyle: {
-    fontSize: 20,
-    fontWeight: "700",
-  },
-  containerStyle: {
-    width: 180,
-  },
-  buttonStyle: {
-    backgroundColor: "rgba(90, 154, 230, 1)",
-    height: 60,
-    borderColor: "transparent",
-    borderWidth: 0,
-    borderRadius: 4,
-  },
-});
 
 type ButtonProps = {
   title: string;
@@ -46,3 +30,20 @@ export const Button: FC<ButtonProps> = ({
     />
   );
 };
+
+const styles = StyleSheet.create({
+  titleStyle: {
+    fontSize: 20,
+    fontWeight: "700",
+  },
+  containerStyle: {
+    width: 180,
+  },
+  buttonStyle: {
+    backgroundColor: "rgba(90, 154, 230, 1)",
+    height: 60,
+    borderColor: "transparent",
+    borderWidth: 0,
+    borderRadius: 4,
+  },
+});
